+++ Psado/Private/New-Header.ps1
@@ -20,7 +20,7 @@ Function New-Header {
     .NOTES
     Private function
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>