+++ Psado/Public/Connect-PSADO.ps1
@@ -24,7 +24,7 @@ function Connect-PSADO {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
 #>