+++ Psado/Public/Get-PSADOProject.ps1
@@ -32,7 +32,7 @@ function Get-PSADOProject {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>