+++ Psado/Public/New-PSADOproject.ps1
@@ -37,7 +37,7 @@ function New-PSADOProject {
 
     .NOTES
     Author: Barbara Forbes
-    Module: PSAzureDevOps
+    Module: Psado
     https://4bes.nl
     @Ba4bes
     #>